% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceSRS.R
\name{infereceSRS}
\alias{infereceSRS}
\title{Outputs a table with SRS based inferences}
\usage{
infereceSRS(sample.data, num.cols = NULL, cat.cols = NULL, alpha,
  N = Inf, type = c("mean", "total"), print.report = FALSE,
  labels = NULL)
}
\arguments{
\item{sample.data}{A data.frame representing a sample with the collected data.}

\item{num.cols}{The number of the columns where the data is continuous.}

\item{cat.cols}{The number of the columns where the data is categorical.}

\item{alpha}{(1 - confidence level).}

\item{type}{If you want means or totals.}

\item{print.report}{\code{FALSE} by default. If \code{TRUE}, prints a .docx table.}

\item{labels}{An optional string vector of labels to be putted into printed table.
Must have the same length of the sum of numerical columns and all levels of categorical columns.}
}
\value{
  A data.frame containing the fields:
  \item{parameter} The label of the parameter.
  \item{n} The effective sample size.
  \item{cv} Variation coefficient
  \item{point.estimate} Point estimate.
  \item{interval} Confidence interval.
}
\description{
This function do Simple Random Sample inferences from a sample data,
given a confidence level. It can calculate estimates for means and totals and proportions too.
It have option to print a Word *.docx table with the estimate parameters.
}
\examples{
data('iris')

infereceSRS(iris,num.cols = 1:4,cat.cols = 5,alpha = 0.9, N=56892,
type = "total")

rotulos <- c("Media da espessura da petala","Media do comprimento da petala",
"Media da espessura da sepala","Media do comprimento da sepala",
"Proporcao de setosas","Proporcao de versicolor", "Proporcao de virginica")
infereceSRS(iris,num.cols = 1:4,cat.cols = 5,alpha = 0.95, N=56892,
             type = "mean",labels = rotulos,print.report = TRUE)
}
