% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Outputs a table with SRS and Stratified based inferences}
\usage{
inference(sample.data, num.y = NULL, denom.x = NULL, total.x = NULL,
  cat.y = NULL, strat = NULL, post.strat = NULL, alpha, N = Inf,
  type = rep("mean", length(num.y) + length(cat.y)), fpc = NULL,
  labels = NULL, print.report = FALSE)
}
\arguments{
\item{sample.data}{A data.frame representing a sample with the collected data.}

\item{num.y}{The index number of the columns (vector) where the data is continuous.}

\item{total.x}{If you are using ratio estimators and want to estimate totals, you gonna need a total of \eqn{\bar{X}}. Use vectors of the same length as your estimated variables for estimate totals. If you don't provide totals, they will be estimated from data.}

\item{cat.y}{The index number of the columns where the data is categorical.}

\item{strat}{The index number of the columns to be used as stratified design.}

\item{post.strat}{The index number of the columns to used as a post stratified design.}

\item{alpha}{1 - Confidence level.}

\item{N}{Population size.}

\item{type}{A vector of characters. Allowed means, totals or ratios. The function use \code{mean} as default.}

\item{fpc}{The index number of the column to be used as the factor of population correction.}

\item{labels}{An optional string vector of labels to be putted into printed table.}

\item{print.report}{\code{FALSE} by default. If \code{TRUE}, prints a .docx table.
Must have the same length of the sum of numerical columns and all levels of categorical columns.}

\item{denom.y}{The index number of the columns (vector) that will be used as denominator for ratio estimators. Use NULL if you don't want to use ratio estimators or combine regular estimators with ratio ones.}
}
\value{
A data.frame containing the following:
\describe{
\item{parameter}{The label of the parameter.}
\item{n}{The effective sample size.}
\item{se}{Standard error.}
\item{point.estimate}{Point estimate.}
\item{interval}{Confidence interval.}
}
}
\description{
This function do Simple Random Sample and Stratified inferences from a sample data,
given an alpha level. It can calculate estimates for means, totals, proportions and ratios too.
It have option to print a Word *.docx table with the estimate parameters and two plotting graph types.
}
