% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCluster.R
\name{nClusterProp}
\alias{nClusterProp}
\title{Calculates the number os cluster to a cluster sample design}
\usage{
nClusterProp(data, clustername, alpha, moe, pq = 0.25)
}
\arguments{
\item{data}{The data.frame containing the variable denoted as the cluster.}

\item{clustername}{The variable name of the considered cluster.}

\item{alpha}{Confidence level.}

\item{moe}{Margin of error.}

\item{pq}{The variance, as \eqn{P(1-P)}. By default, it is 0.25.}
}
\value{
The value (integer) of the number of clusters to sample in cluster sampling.
}
\description{
This function returns the number of clusters to be sampled in a proportion based survey.
By default, the argument `pq` is 0.25 (maximum variance).
}
\examples{
library(survey)
data("api")
nClusterProp(data=apipop, clustername = "dnum", alpha = 0.95, moe = 0.1)
nClusterProp(data=apipop, clustername = "dnum", alpha = 0.95, moe = 0.05)
}
