% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neymanAlloc.R
\name{neymanAllocation}
\alias{neymanAllocation}
\title{Returns the number of elements per strata based on Neyman optimum allocation}
\usage{
neymanAllocation(n, strat.vector = NULL, sample.vector, sd.vector)
}
\arguments{
\item{n}{The size of sample.}

\item{strat.vector}{The vector of strata from population data.}

\item{sample.vector}{The vector of strata from sample data. Required.}

\item{sd.vector}{The data sample vector from which the standards deviations will be calculated.}
}
\value{
Vector with the same length as the unique labels of strata containing the size of each strata for sampling.
}
\description{
Based on data from sample and population, this function returns the optimum allocation of elements.
Users can input strata vectors from sample and from population, but the last is not required.
The output is a vector.
}
