% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPlan.R
\name{nPlan}
\alias{nPlan}
\title{Returns a data.frame with total resources to be spend in sampling process, based on the thresholds defined}
\usage{
nPlan(hh, max.hh = NULL, max.n = NULL, max.alpha = NULL,
  max.beta = NULL, fun = "nSRS", ...)
}
\arguments{
\item{hh}{The cost, measured in time (HH), for collecting data for each sample unit.}

\item{max.hh}{The cost threshold from which results and their parameters will be excluded.}

\item{max.n}{The sample size threshold from which results and their parameters will be excluded.}

\item{max.alpha}{The maximum z-value of \eqn{\alpha} from which results and their parameters will be excluded.}

\item{max.beta}{The maximum z-value of \eqn{\beta} from which results and their parameters will be excluded. Applicable only to \code{nPowerSample} function.}

\item{fun}{Function used to calculate samplesize. Can be \code{nSRS}, \code{nStrata} or \code{nSampleSize}.}

\item{...}{Others parameters passed to be used with the sample size functions. See examples.}
}
\value{
A data.frame with
\describe{
\item{alpha}{The \eqn{\alpha}.}
\item{moe}{The margin of error, when using \code{nSRS} or \code{nStrata}.}
\item{beta}{The \eqn{\alpha}, when using \code{nSampleSize}.}
\item{n}{Respective sample size.}
\item{hh}{Cost, in time, as the argument passed to the function.}
\item{hh_total}{Total cost, in time.}
}
}
\description{
This function returns a data.frame with the calculations of time to be spend on sampling process, listing respectives sampling parameters, like confidence levels and margin of errors.
Need a sample size function to work.
}
\examples{
################### nSRS
nPlan(hh = (10/60), fun = "nSRS", max.alpha = 0.1, max.n = 176, max.hh = 60,
      moe = seq(0.01,0.1,by=0.01), pq = 0.25, N = 16548)

################### nPowerSample
nPlan(hh = (5/60), fun = "nPowerSample", max.hh = 40, max.alpha = 0.1, max.beta = 0.2,
      p=0.02, p0=0.05, alternative = 1)
}
