% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPowerSample.R
\name{nPowerSample}
\alias{nPowerSample}
\title{Calculates the sample size for hypotesis tests}
\usage{
nPowerSample(p = NULL, p0 = NULL, mu = NULL, mu0 = NULL,
  sd = NULL, alpha = 0.05, beta = 0.2, alternative = 1)
}
\arguments{
\item{p}{Population proportion.}

\item{p0}{Hypothesized proportion.}

\item{mu}{Population mean.}

\item{mu0}{Hypothesized mean.}

\item{alpha}{Level of significance.}

\item{alternative}{\code{1} if is a greater/lesser test and \code{2} for equal/different one.}
}
\value{
The sample size.
}
\description{
This function calculates de sample size for hypotesys tests considering the following hypotesis:
\deqn{H_{0}:\text{ } \mu \neq \mu_{0}}
\deqn{H_{0}:\text{ } \bar{p} \neq p_{0}}
\deqn{H_{0}:\text{ } \mu \leq \text{ or } \geq \mu_{0}}
\deqn{H_{0}:\text{ } \bar{p} \leq \text{ or } \geq p_{0}}
}
\examples{
nPowerSample(p=0.03, p0=0.05)
nPowerSample(p=0.02, p0=0.05)
nPowerSample(mu=100, mu0=120,sd=24)
nPowerSample(mu=100, mu0=110,sd=24,beta=0.05,alternative = 2)
}
