% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSRS.R
\name{nSRS}
\alias{nSRS}
\title{Calculates the sample size for SRS}
\usage{
nSRS(moe, alpha, CV = NULL, S2 = NULL, pq = NULL, N = Inf)
}
\arguments{
\item{moe}{Margin of error.}

\item{alpha}{1 - (confidence level).}

\item{CV}{Variaton coefficient.}

\item{S2}{\eqn{\sigma^2}, population variance.}

\item{pq}{\eqn{P(1 - P)}, population variance for proportions.}

\item{N}{Population size.}
}
\value{
The value of \eqn{n}.
}
\description{
This function calculates de sample size in a SRS design based on one of three possible
arguments: \eqn{\sigma^2}, \eqn{pq} or \eqn{CV}. Provide just one of then.
}
\examples{
nSRS(moe=0.05,alpha=0.05,S2=6,N=1500)
nSRS(moe=0.05,alpha=0.05,CV=0.2,N=1500)
nSRS(moe=0.05,alpha=0.05,pq=0.25,N=1500)
}
