% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nStrata.R
\name{nStrata}
\alias{nStrata}
\title{Calculates the sample size in a stratified desgin}
\usage{
nStrata(data, stratanames, alpha, moe, S2 = NULL, pq = NULL,
  V = NULL, N = Inf, method = c("none", "prop", "optimum"))
}
\arguments{
\item{data}{The data.frame containing the variable denoted as the cluster.}

\item{stratanames}{The variable name of the considered strata. Can be a vector of variable names.}

\item{alpha}{Confidence level.}

\item{moe}{Margin of error.}

\item{S2}{The variance, as \eqn{\sigma^2}. Not necessary if `pq` or `V` is not `NULL`.}

\item{pq}{The variance, as \eqn{P(1-P)}. Not necessary if `S2` or `V` is not `NULL`.}

\item{V}{The variance, as \eqn{(d/t)^2}. Not necessary if `S2` or `pq` is not `NULL`.}

\item{N}{Population size.}

\item{method}{A string with `none`, `prop` or `optimum`. With `optimum`, it uses Neyman allocation.}
}
\value{
The value (integer) of the number of elements to sample in stratified sampling. If `prop` or `optimum`
where used, the function returns a list with the components
\describe{
\item{n}{Total number of elements in the sample.}
\item{nh}{In case of method `prop` or `optimum`, the allocation in each strata.}
}
}
\description{
This function returns the number of elements to be sampled in a stratified based survey.
If the method choosen is `prop` or `optimum`, the function return the allocation too.
}
\examples{
library(survey)
data("api")
nStrata(apipop,stratanames = "stype", alpha = 0.95, moe = 0.05,
        pq = 0.25, N = Inf, method = "none")
nStrata(apipop,stratanames = "stype", alpha = 0.95, moe = 0.05,
        pq = 0.25, N = nrow(apipop), method = "prop")
nStrata(apipop,stratanames = "stype", alpha = 0.95, moe = 0.05,
        pq = 0.25, N = nrow(apipop), method = "optimum")
### Cochran's example (Cochran, W. G. (2007). Sampling techniques. John Wiley & Sons. p. 106-107)
data_cochran=data.frame(stratum=rep(c(1:6),c(13,18,26,42,73,24)))
nStrata(data_cochran,stratanames = "stratum", N=196, method = "optimum",
S2=c(325^2,190^2,189^2,82^2,86^2,190^2),V=7974976)
}
