% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSubpop.R
\name{NSubpop}
\alias{NSubpop}
\title{Calculates the new subpopulation due elements that not belong to the populational universe}
\usage{
NSubpop(N, n, p, z = 3)
}
\arguments{
\item{N}{Population size.}

\item{n}{Sample size.}

\item{p}{Proportion or the number of elements that belong to the universe intended.}

\item{z}{The z-value of \eqn{\alpha}.}
}
\value{
The value (integer) of the new population (subpopulation) number.
}
\description{
This function returns the number of the new subpopulation.
}
\examples{
NSubpop(45000,100,59)
NSubpop(45000,100,5)
}
