% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Calculates the statistics for hypotesis tests}
\usage{
test(sample.data, num.col = NULL, prop.col = NULL, mu0 = NULL,
  p0 = NULL, alpha, alternative = NULL, labels = NULL,
  plot.graph = FALSE)
}
\arguments{
\item{sample.data}{The source \code{data.frame} from which the funtion will search the columns.}

\item{num.col}{Vector of the numeric column numbers.}

\item{prop.col}{Vector of the binary column numbers.}

\item{mu0}{Hypothesized mean.}

\item{p0}{Hypothesized proportion.}

\item{alpha}{Level of significance.}

\item{alternative}{\code{1} if is a greater/lesser test and \code{2} for equal/different one.}

\item{labels}{Label vector for the outputs.}

\item{plot.graph}{Plot a confidence intervals for proportions only. Not implemented yet.}
}
\value{
\code{data.frame} containing the following:
\describe{
\item{var}{The label of the variable/parameter.}
\item{point}{Point estimate.}
\item{tolerance}{The hypothesized mean/proportion.}
\item{sdd_err}{Standard error.}
\item{z_value}{Statistics z.}
\item{pval}{P-value.}
\item{power}{Test power, defined as (\eqn{1 - \beta}).}
}
}
\description{
This function calculates de statistics for hypotesys tests considering the following hypotesis:
\deqn{H_{0}:\text{ } \mu \neq \mu_{0}}
\deqn{H_{0}:\text{ } \bar{p} \neq p_{0}}
\deqn{H_{0}:\text{ } \mu \leq \text{ or } \geq \mu_{0}}
\deqn{H_{0}:\text{ } \bar{p} \leq \text{ or } \geq p_{0}}
}
\examples{
n=168
set.seed(101)
dados <- data.frame(CONTROLE1=rbinom(n,1,prob = 0.4),
                    CONTROLE2=rbinom(n,1,prob = 0.02),
                    CONTROLE3=rbinom(n,1,prob = 0.1),
                    CONTROLE4=rbinom(n,1,prob = 0.01),
                    CONTROLE5=rbinom(n,1,prob = 0.2),
                    CONTROLE6=rnorm(n,10,3),
                    CONTROLE7=rnorm(n,250,20))
test(sample.data = dados, num.col = 6:7, prop.col = 1:5, alpha = 0.05,
     mu0 = c(10,100),p0=c(0.05,0.05,0.10,0.10,0.1),
     alternative = c(rep(1,5),rep(2,2)))
}
