% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSample.R
\name{selectSample}
\alias{selectSample}
\title{Select sample and returns the \code{data.frame}}
\usage{
selectSample(data, stratum = NULL, method, size, seed,
  resample.data = NULL)
}
\arguments{
\item{data}{The data.frame containing the population data.}

\item{stratum}{A vector with strata variable names.}

\item{method}{Select between \code{srswr}, \code{srswor} or \code{systematic}.}

\item{size}{The size of sample. Must provide a vector if the sample is stratified.}

\item{seed}{The seed for the random generator.}

\item{resample.data}{A data frame used as pilot sample. Its useful to ignore the observations sampled before and to maintain the original fpc.}
}
\value{
A \code{data.frame} additionally with \code{Prob} and \code{fpc} columns. If the systematic
method of selection has been chosen, returns the \code{k} and cumulated \code{k} also.
}
\description{
Sample and outputs a data.frame with desired technique and size.
}
